feat: refactor screenshot handling and implement cleanupOldFiles utility for improved file management
